Fixed signed / unsigned comparison issue in PlatformIO test.

+++ test/PlatformIOTest.cpp
@@ -101,7 +101,7 @@ class PlatformIOTest : public ::testing::Test
         std::list<std::shared_ptr<PlatformIOTestMockIOGroup> > m_iogroup_ptr;
         std::unique_ptr<PlatformIO> m_platio;
         MockPlatformTopo m_topo;
-        const int M_NUM_CPU = 4;
+        const unsigned int M_NUM_CPU = 4;
 };
 
 void PlatformIOTest::SetUp()
@@ -316,7 +316,7 @@ TEST_F(PlatformIOTest, push_signal_agg)
     EXPECT_EQ(0, m_platio->num_signal());
     // Domain of FREQ is CPU
     m_platio->push_signal("FREQ", IPlatformTopo::M_DOMAIN_PACKAGE, 0);
-    EXPECT_EQ(1 + M_NUM_CPU, m_platio->num_signal());
+    EXPECT_EQ(1 + M_NUM_CPU, (unsigned int)m_platio->num_signal());
 }
 
 TEST_F(PlatformIOTest, signal_power)
@@ -433,7 +433,7 @@ TEST_F(PlatformIOTest, push_control_agg)
     EXPECT_EQ(0, m_platio->num_control());
 
     m_platio->push_control("FREQ", IPlatformTopo::M_DOMAIN_PACKAGE, 0);
-    EXPECT_EQ(1 + M_NUM_CPU, m_platio->num_control());
+    EXPECT_EQ(1 + M_NUM_CPU, (unsigned int)m_platio->num_control());
 }
 
 TEST_F(PlatformIOTest, sample)
@@ -481,7 +481,7 @@ TEST_F(PlatformIOTest, sample_agg)
             EXPECT_CALL(*it, signal_domain_type("FREQ")).Times(2 + M_NUM_CPU);
             EXPECT_CALL(*it, agg_function("FREQ"))
                 .WillOnce(Return(geopm::Agg::average));
-            for (int ii = 0; ii < M_NUM_CPU; ++ii) {
+            for (unsigned int ii = 0; ii < M_NUM_CPU; ++ii) {
                 EXPECT_CALL(*it, push_signal("FREQ", IPlatformTopo::M_DOMAIN_CPU, ii))
                     .WillOnce(Return(ii));
                 EXPECT_CALL(*it, sample(ii)).WillOnce(Return(ii));
@@ -490,7 +490,7 @@ TEST_F(PlatformIOTest, sample_agg)
         EXPECT_CALL(*it, read_batch());
     }
     double sum = 0;
-    for (int ii = 0; ii < M_NUM_CPU; ++ii) {
+    for (unsigned int ii = 0; ii < M_NUM_CPU; ++ii) {
         sum += ii;
     }
     int freq_idx = m_platio->push_signal("FREQ", IPlatformTopo::M_DOMAIN_PACKAGE, 0);
@@ -534,7 +534,7 @@ TEST_F(PlatformIOTest, adjust_agg)
             // called once for initial push, once for converting domain, once for each cpu,
             EXPECT_CALL(*it, control_domain_type("FREQ")).Times(2 + M_NUM_CPU);
             // adjusting package will adjust every cpu on the package
-            for (int ii = 0; ii < M_NUM_CPU; ++ii) {
+            for (unsigned int ii = 0; ii < M_NUM_CPU; ++ii) {
                 EXPECT_CALL(*it, push_control("FREQ", IPlatformTopo::M_DOMAIN_CPU, ii))
                     .WillOnce(Return(ii));
                 EXPECT_CALL(*it, adjust(ii, value));
